+++ Holovibes/sources/thread/file_frame_read_worker.cc
@@ -20,7 +20,7 @@
 namespace holovibes::worker
 {
     FileFrameReadWorker::FileFrameReadWorker(const std::string& file_path, bool loop, unsigned int fps,
-                                            size_t first_frame_id, size_t total_nb_frames_to_read,
+                                            unsigned int first_frame_id, unsigned int total_nb_frames_to_read,
                                             bool load_file_in_gpu,
                                             std::atomic<std::shared_ptr<Queue>>& gpu_input_queue) :
         FrameReadWorker(gpu_input_queue),
